requirements.txt
Adafruit-Blinka==8.51.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-neopixel==6.3.14
adafruit-circuitpython-pixelbuf==2.0.7
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
mfrc522==0.0.7
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.3.0
